Move FetchMySQLVersion to mysql/protocol package.
